New loop method, updated readme and bugfixes
New loop method for datestatustimer causing it to no longer need force,
except when there are changes to the date or date name, which should be
fixed soon.
Fixed a bug where it wouldn't let you set the date name to null.
Made spacing in isitchristmas consistent among all functions.
Updated readme to reflect the changes.